+++ game/src/ui/text_data.h
@@ -2,7 +2,8 @@
 #define TEXT_DATA_H
 
 extern const unsigned char s_controls_data[274];
-extern const unsigned char s_credits_data[560];
+extern const unsigned char s_credits_data[598];
+extern const unsigned char s_story_data[1116];
 
 void RenderTextScreen(const unsigned char* data, int len, const char* backHint);